feat: migrate build system to Rolldown and improve TypeScript configuration
- Replace Rollup + ESBuild with Rolldown for faster, simpler builds
- Remove redundant type declarations, leverage TypeScript inference
- Modernize TypeScript config (ES2020 target, bundler resolution)
- Simplify build scripts and configuration files
- Add ES module support with "type": "module"
- Reduce build dependencies from 7 packages to just Rolldown

Build performance: ~90ms with Rolldown (native TypeScript support)
All 70 tests passing successfully
