+++ package.json
@@ -40,15 +40,14 @@
   "main": "lib/index.js",
   "module": "lib/index.es.js",
   "types": "lib/index.d.ts",
+  "type": "module",
   "files": [
     "lib",
     "resources"
   ],
   "scripts": {
-    "build": "yarn build:main && yarn build:serverless",
-    "build:main": "rm -rf lib && rollup -c rollup.config.cjs && yarn build:types",
-    "build:serverless": "rollup -c rollup.config.serverless.cjs",
-    "build:types": "tsc --declaration --emitDeclarationOnly --outDir lib --declarationMap",
+    "build": "rm -rf lib && rolldown -c && yarn build:types",
+    "build:types": "tsc --emitDeclarationOnly",
     "check": "biome check .",
     "format": "biome format --write .",
     "lint": "biome lint .",
@@ -58,9 +57,8 @@
     "prepublishOnly": "yarn build",
     "release": "semantic-release",
     "release:dry": "semantic-release --dry-run",
-    "test": "jest",
-    "watch": "rm -rf lib && rollup -cw rollup.config.cjs",
-    "watch:serverless": "rollup -cw rollup.config.serverless.cjs"
+    "test": "NODE_OPTIONS='--experimental-vm-modules' jest",
+    "watch": "rm -rf lib && rolldown -c --watch"
   },
   "config": {
     "commitizen": {
@@ -83,21 +81,15 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.2.2",
-    "@rollup/plugin-commonjs": "^28.0.6",
-    "@rollup/plugin-json": "^6.1.0",
-    "@rollup/plugin-node-resolve": "^16.0.1",
-    "@rollup/plugin-terser": "^0.4.4",
     "@types/bson": "^4.2.4",
     "@types/jest": "^30.0.0",
     "@types/node": "^24.3.0",
     "@types/shelljs": "^0.8.17",
-    "esbuild": "^0.25.9",
     "husky": "^9.1.7",
     "jest": "^30.1.3",
     "lint-staged": "^16.1.6",
     "prettier": "^3.6.2",
-    "rollup": "^4.50.0",
-    "rollup-plugin-esbuild": "^6.2.1",
+    "rolldown": "^1.0.0-beta.34",
     "shelljs": "^0.10.0",
     "ts-jest": "^29.4.1",
     "tslib": "^2.8.1",
