+++ tsconfig.json
@@ -1,16 +1,19 @@
 {
   "compilerOptions": {
-    "target": "ES5",
-    "module": "esnext",
-    "lib": ["ESNext"],
+    "target": "ES2020",
+    "module": "ESNext",
+    "lib": ["ES2020"],
     "declaration": true,
+    "declarationMap": true,
+    "outDir": "./lib",
+    "rootDir": "./src",
     "strict": true,
-    "noImplicitAny": true,
-    "moduleResolution": "node",
-    "types": ["node", "jest"],
+    "moduleResolution": "bundler",
     "esModuleInterop": true,
-    "downlevelIteration": true
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "resolveJsonModule": true
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "**/*.test.ts"]
+  "exclude": ["node_modules", "**/*.test.ts", "**/*.spec.ts"]
 }
